% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyttest.R
\name{applyttest}
\alias{applyttest}
\title{An APAF wrapper for t.test() that works for both proteins and peptides}
\usage{
applyttest(
  data,
  Group,
  PeporProt = c("Peptide", "Protein"),
  doLogs = TRUE,
  numerator = levels(Group)[1]
)
}
\arguments{
\item{data}{input matrix, is as subset of either prot.ag or pep.ag}

\item{Group}{Group for the labels}

\item{PeporProt}{Specify 'Protein' or 'Peptide' depending on which type of matrix is being used}

\item{doLogs}{take the log() of the matrix. Defaults to TRUE}

\item{numerator}{the Group for the comparison, defaults to levels(Group[1])}
}
\description{
An APAF wrapper for t.test() that works for both proteins and peptides
}
