% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{TO DO: make a clean function that can clean data based on its type. Blank for now
--clean argument should have a flag for the search engine used, e.g. MaxQuant or Protein Discoverer. Version numbers matter as well.}
\usage{
clean(exp, filelist, data)
}
\arguments{
\item{data}{}
}
\description{
TO DO: make a clean function that can clean data based on its type. Blank for now
--clean argument should have a flag for the search engine used, e.g. MaxQuant or Protein Discoverer. Version numbers matter as well.
}
