% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClusterDendrogram.R
\name{createClusterDendrogram}
\alias{createClusterDendrogram}
\title{Makes CLUST1Gene dendrogram, but also ClusterExpressionProfiles? TO DO: sort out who does what between this and HClust()}
\usage{
createClusterDendrogram(
  res1,
  cluster.data,
  clusterMetric,
  Cluster,
  Anova.idx,
  Group
)
}
\arguments{
\item{res1}{Dunno, this should be renamed here and in the main script}

\item{Group}{}
}
\description{
Makes CLUST1Gene dendrogram, but also ClusterExpressionProfiles? TO DO: sort out who does what between this and HClust()
}
