% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createResultsImagesxlsx.R
\name{createResultsImagexlsx}
\alias{createResultsImagexlsx}
\title{Creates an Excel workbook with each image and a description. Doesn't require any arguments -- creating this workbook is self-contained}
\usage{
createResultsImagexlsx()
}
\description{
Creates an Excel workbook with each image and a description. Doesn't require any arguments -- creating this workbook is self-contained
}
