% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createResultsxlsx.R
\name{createResultsxlsx}
\alias{createResultsxlsx}
\title{Creates an Excel workbook all results. Later it should format based upon experiment. TO DO: parameter values like normalization etc should be passed in}
\usage{
createResultsxlsx(exp = c("SWATH, TMT"), list = inputlist)
}
\arguments{
\item{exp}{Either SWATH or TMT}

\item{list}{A list of data frames containing data for excel output tabs}
}
\description{
Creates an Excel workbook all results. Later it should format based upon experiment. TO DO: parameter values like normalization etc should be passed in
}
