% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniprotBatch.R
\name{getUniprotBatch}
\alias{getUniprotBatch}
\title{Get UniProt annotations}
\usage{
getUniprotBatch(
  values = IDList,
  attributes = c("accession", "protein_name", "go_id")
)
}
\arguments{
\item{values}{IDList, ideally as sp|UniprotID|organism}

\item{attributes}{UniProt information to obtain. Default obtains accession, protein name and GO ID}
}
\description{
Get UniProt annotations
}
