% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergebyUniprot.R
\name{mergebyUniprot}
\alias{mergebyUniprot}
\title{Merge the Uniprot annotation with the results}
\usage{
mergebyUniprot(input_table, ANOVA_TABLE = FALSE)
}
\arguments{
\item{ANOVA_TABLE}{TRUE or FALSE whether merging in the ANOVA results}
}
\value{

}
\description{
Merge the Uniprot annotation with the results
}
