% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrGroup.R
\name{mlrGroup}
\alias{mlrGroup}
\title{Multiple linear regression that uses Groups. No plotting}
\usage{
mlrGroup(mat, Group)
}
\arguments{
\item{mat}{matrix of ion intensities}

\item{Group}{Group from the design file}
}
\description{
Multiple linear regression that uses Groups. No plotting
}
