% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{Any normalisation of the data matrix applied to APAF data}
\usage{
normalise(data_raw, normalisation, Group)
}
\arguments{
\item{data_raw}{The un-normalised raw ion data matrix, as a list}

\item{normalisation}{comes from the --normalisation flag and defaults to "total"}

\item{Group}{}
}
\value{

}
\description{
Any normalisation of the data matrix applied to APAF data
}
