% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Any normalisation of the data matrix applied to APAF data}
\usage{
normalize(exp, data, normalization, Group)
}
\arguments{
\item{normalization}{comes from the --normalization flag and decfults to "total"}

\item{Group}{}

\item{data_raw}{The un-normalised raw ion data matrix, as a list}
}
\value{
data_norm the normalised data. The type of normalisation depends on the experiment and the normalization environment variable set through --normalization
}
\description{
Any normalisation of the data matrix applied to APAF data
}
