% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_calc.R
\name{PCA_calc}
\alias{PCA_calc}
\title{Compute PCA, no plots}
\usage{
PCA_calc(
  experiment = c("SWATH", "TMT"),
  data,
  scaleR = FALSE,
  scaleC = TRUE,
  k = min(dim(data)) - 2
)
}
\arguments{
\item{experiment}{experiment: either SWATH or TMT}

\item{data}{The prot.ag dataframe. Long format intensity? data with protein name as first column.}

\item{scaleR}{TO BE DETERMINED - leave as FALSE as on GenePattern - manual row scaling?}

\item{scaleC}{Pass scaling argument to prcomp function for PCA. Default TRUE.}

\item{k}{Number of components to compute. Defaults to number of variables - 1.}
}
\value{
The components of the PCA that are labelled for easy plotting
}
\description{
Compute PCA, no plots
}
