% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plot.R
\name{PCA_plot}
\alias{PCA_plot}
\title{Generate 2D and 3D PCA plots}
\usage{
PCA_plot(exp = c("SWATH", "TMT"), data, pca.components, Group)
}
\arguments{
\item{exp}{Currently TMT or SWATH}

\item{data}{Currently only used by TMT to get rownames}

\item{pca.components}{The output of PCA_calc()}

\item{Group}{Group information for each column from PCA_calc() input (usually prot.ag). No replicate information.}
}
\description{
Generate 2D and 3D PCA plots
}
