% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoxPlots.R
\name{plotBoxPlots}
\alias{plotBoxPlots}
\title{Plotting any APAF boxplots. To be done using ggplot, and take in any experiment type}
\usage{
plotBoxPlots(
  experiment = c("SWATH", "TMT"),
  data,
  normalisation,
  design,
  Group
)
}
\arguments{
\item{experiment}{the experiment type}

\item{data}{for TMT this is the ion data matrix}

\item{normalisation}{The types of normalisation from APAFunctions::normalise() should  be listed here, but those are yet to be finalised}

\item{design}{the processed design file object that contains labels}

\item{Group}{groups that is represented as factors}
}
\description{
Plotting any APAF boxplots. To be done using ggplot, and take in any experiment type
}
