% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterDendrogram.R
\name{plotClusterDendrogram}
\alias{plotClusterDendrogram}
\title{The one function to plot dendrograms. Future: maybe rewrite with ggdendro?}
\usage{
plotClusterDendrogram(
  experiment = c("SWATH", "TMT"),
  hclust_result,
  cluster.data,
  Group,
  clusterMetric,
  Cluster,
  Anova.idx,
  glabel,
  clabel = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
The one function to plot dendrograms. Future: maybe rewrite with ggdendro?
}
