% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotColouredDendrogram.R
\name{plotColouredDendrogram}
\alias{plotColouredDendrogram}
\title{plot a dendrogram, normally called after HClust()}
\usage{
plotColouredDendrogram(result, glabel, clabel = NULL, ...)
}
\arguments{
\item{result}{The results of the HClust() function}

\item{glabel}{Group labels}

\item{clabel}{column? cluster? labels TO DO: check this}

\item{...}{Ellipses allow extra arguments to the function, taken in as a list}
}
\description{
plot a dendrogram, normally called after HClust()
}
