% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelationPlots.R
\name{plotCorrelationPlots}
\alias{plotCorrelationPlots}
\title{Generic correlation plot function}
\usage{
plotCorrelationPlots(data, Group)
}
\arguments{
\item{Group}{}
}
\description{
Generic correlation plot function
}
