% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotErrorBarsLines.R
\name{plotErrorBarsLines}
\alias{plotErrorBarsLines}
\title{plot error bars with lines included}
\usage{
plotErrorBarsLines(
  v,
  barSizes,
  lines,
  labels = NULL,
  color = "blue",
  ylim = c(min(lines), max(lines)),
  ...
)
}
\arguments{
\item{v}{Matrix of aggregated genes}

\item{barSizes}{error bar sizes, normall twice the standard deviation}

\item{lines}{lines from sample cluster data}

\item{labels}{labels normal provided from ordered group, but defaults to NULL}

\item{color}{colours, often provided by the rainbow() function}

\item{ylim}{y limit of the plot, the normally the minimum value}

\item{...}{ellipses allow extra arguments to a function, when they are provided as a list}
}
\description{
plot error bars with lines included
}
