% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRelativeDensities.R
\name{plotRelativeDensities}
\alias{plotRelativeDensities}
\title{Plot the relative of ions (provide ion matrix w/ or wo/ normalisation)}
\usage{
plotRelativeDensities(data, Group = NULL, idx = NULL, main = "Densities")
}
\arguments{
\item{data}{matrix of ion intensities}

\item{Group}{Group from the design file}

\item{idx}{index of the group?}

\item{main}{The label? of the densities}
}
\description{
Plot the relative of ions (provide ion matrix w/ or wo/ normalisation)
}
