% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printxlsxExplanationSheet.R
\name{printxlsExplanationSheet}
\alias{printxlsExplanationSheet}
\title{Print xlsx explanation page information key}
\usage{
printxlsExplanationSheet(wb, tabName, file)
}
\arguments{
\item{wb}{Keep the same workbook to maintain the styles}

\item{tabName}{One of 'Key', 'Design', 'Comparisons', or 'InputParameters'}

\item{file}{The file associated with the output for the tab. If tabName = 'Design', design. If tabName = 'Comparisons', comps. If tabName = 'InputParameters', params. Not required for 'Key'.}
}
\description{
Print xlsx explanation page information key
}
