% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printxlsxDataSheet.R
\name{printxlsxDataSheet}
\alias{printxlsxDataSheet}
\title{Set openxlsx style for comparisons and ANOVA table}
\usage{
printxlsxDataSheet(
  wb,
  data,
  FCcol,
  pvalcol,
  tabName = "results",
  hicutoff = 1.5,
  lowcutoff = 0.67,
  pvalcutoff = 0.05
)
}
\arguments{
\item{wb}{Keep the same workbook to maintain the styles}

\item{data}{If more than sheet to be added, needs to be in for loop}

\item{FCcol}{Column index for Fold Change values}

\item{pvalcol}{Column index for pvalues}

\item{tabName}{Tab name, specify each time to avoid overwriting previous tabs}

\item{hicutoff}{Default 1.5}

\item{lowcutoff}{Default 0.67}

\item{pvalcutoff}{Default 0.05}
}
\description{
Set openxlsx style for comparisons and ANOVA table
}
