% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printxlsxInfoSheet.R
\name{printxlsxInfoSheet}
\alias{printxlsxInfoSheet}
\title{Print xlsx information sheet}
\usage{
printxlsxInfoSheet(wb, tabName, file)
}
\arguments{
\item{wb}{Keep the same workbook to maintain the styles}

\item{tabName}{Name of tab for results file. 'Key' will insert the SWATH column fields.}

\item{file}{The file associated with the input data for the tab. Not required for 'Key'.}
}
\description{
Print xlsx information sheet
}
