% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInputFiles.R
\name{readInputFiles}
\alias{readInputFiles}
\title{Read mass spec input files and design files for analysis and writes outputs to global environment}
\usage{
readInputFiles(filename, designfile)
}
\arguments{
\item{filename}{Path to data file}

\item{designfile}{Path to design file}
}
\value{

}
\description{
Read mass spec input files and design files for analysis and writes outputs to global environment
}
