% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMT_normalisation.R
\name{TMT_normalisation}
\alias{TMT_normalisation}
\title{TMT normalisation containing both SampleLoadNorm and IRS. To be merged eventually}
\usage{
TMT_normalisation(data, SampleLoadNorm)
}
\arguments{
\item{SampleLoadNorm}{comes from the --SampleLoadNorm flag and decfaults to "total"}

\item{data_raw}{The un-normalised raw ion data matrix, as a list}
}
\value{
data_ls and data_irs in a list. This used for all later ion data in TMT at the moment. It shouldn't be called data_irs. It should be data_processed
}
\description{
TMT normalisation containing both SampleLoadNorm and IRS. To be merged eventually
}
