% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalAreaNorm.R
\name{totalAreaNorm}
\alias{totalAreaNorm}
\title{normalise ion intensities by total area
NB: this isn't used anywhere else, maybe put back into SWATHANOVA.R}
\usage{
totalAreaNorm(mat)
}
\arguments{
\item{mat}{Matrix of ion intensities}
}
\description{
normalise ion intensities by total area
NB: this isn't used anywhere else, maybe put back into SWATHANOVA.R
}
