% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createResultsxlsx.R
\name{createResultsxlsx}
\alias{createResultsxlsx}
\title{Creates an Excel workbook all results. Later it should format based upon experiment. TO DO: parameter values like normalization etc should be passed in}
\usage{
createResultsxlsx(res.list, Anova.table, design, comps)
}
\arguments{
\item{res.list}{List containing every results dataframe}

\item{Anova.table}{Dataframe that contains the ANOVA results}

\item{design}{Experimental design file}

\item{comps}{Comparison between groups}
}
\description{
Creates an Excel workbook all results. Later it should format based upon experiment. TO DO: parameter values like normalization etc should be passed in
}
