% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HClust.R
\name{HClust}
\alias{HClust}
\title{Clustering method. If plot == TRUE, it plots using APAFunctions::plotColouredDendrogram()}
\usage{
HClust(
  exp = c("SWATH", "TMT"),
  data,
  metric = c("euclidean", "manhattan", "pearsonCorrelation"),
  scale = FALSE,
  method = c("single", "complete", "average"),
  glabel = row.names(data),
  clabel = NULL,
  cutNumber = NULL,
  cutHeight = NULL,
  plot = TRUE
)
}
\arguments{
\item{plot}{uses APAFunctions::plotColouredDendrogram(), defaults to TRUE}

\item{basefile}{the base name for the type of hclust method used: e.g. hclust-pearsonCorrelation-complete}
}
\description{
Clustering method. If plot == TRUE, it plots using APAFunctions::plotColouredDendrogram()
}
