% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HClust.R
\name{HClust}
\alias{HClust}
\title{HClust}
\usage{
HClust(
  data,
  metric = c("euclidean", "manhattan", "pearsonCorrelation"),
  scale = FALSE,
  method = c("single", "complete", "average"),
  glabel = row.names(data),
  cutNumber = NULL,
  cutHeight = NULL,
  plot = TRUE
)
}
\arguments{
\item{plot}{uses APAFunctions::plotClusterDendrogram(), defaults to TRUE}
}
\description{
Performs hierarchical clustering to cluster results. Outputs result in list along with method used.
}
