% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HClust.R
\name{HClust}
\alias{HClust}
\title{Clustering method. TO DO -- document later}
\usage{
HClust(
  data = data,
  metric = c("euclidean", "manhattan", "pearsonCorrelation"),
  scale = FALSE,
  method = c("single", "complete", "average"),
  basefile = "hclust",
  glabel = row.names(data),
  clabel = NULL,
  cutNumber = NULL,
  cutHeight = NULL,
  plot = TRUE
)
}
\arguments{
\item{plot}{#plotting should be pulled out. plotting is provided by plotColouredDendrogram()}
}
\description{
Clustering method. TO DO -- document later
}
