% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Any normalisation of the data matrix applied to APAF data}
\usage{
normalize(exp = c("SWATH", "TMT"), data_raw, normalization, Group)
}
\arguments{
\item{data_raw}{The un-normalized raw ion data matrix, as a list}

\item{normalization}{comes from the --normalization flag and defaults to "total"}

\item{Group}{}
}
\value{

}
\description{
Any normalisation of the data matrix applied to APAF data
}
