% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plot.R
\name{PCA_plot}
\alias{PCA_plot}
\title{Generate 2D and 3D PCA plots}
\usage{
PCA_plot(components, labelValue)
}
\arguments{
\item{components}{The output of PCA_calc()}

\item{labelValue}{Group information for each column from PCA_calc() input (usually prot.ag). No replicate information.}
}
\description{
Generate 2D and 3D PCA plots
}
