% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoxPlots.R
\name{plotBoxPlots}
\alias{plotBoxPlots}
\title{Plotting any APAF boxplots. To be done using ggplot, and take in any experiment type}
\usage{
plotBoxPlots(exp, data, design, Group)
}
\arguments{
\item{exp}{the experiment type}

\item{data}{for TMT this is the ion data matrix}

\item{design}{the processed design file object that contains labels}

\item{Group}{groups that is represented as factors}
}
\description{
Plotting any APAF boxplots. To be done using ggplot, and take in any experiment type
}
