% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterProfile.R
\name{plotClusterProfile}
\alias{plotClusterProfile}
\title{plotClusterProfile}
\usage{
plotClusterProfile(cluster.data, clustID, Group, k = 4, ylab = "Abundance")
}
\arguments{
\item{ylab}{}
}
\description{
plotClusterProfile
}
