% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmaps.R
\name{plotHeatmaps}
\alias{plotHeatmaps}
\title{Plots a heatmap with either euclidean or cor distance; ANOVA or no ANOVA}
\usage{
plotHeatmaps(prot.ag, Group, dist, Anova.idx = NULL, useAnova = FALSE)
}
\arguments{
\item{prot.ag}{table with Protein name and intensities by group}

\item{Group}{Group for the labels}

\item{dist}{either [euclidean or cordist]}

\item{Anova.idx}{index of where significant ANOVA values are}

\item{useAnova}{whether ANOVA has been preformed}
}
\description{
Plots a heatmap with either euclidean or cor distance; ANOVA or no ANOVA
}
