% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcano.R
\name{plotVolcano}
\alias{plotVolcano}
\title{The Volcano plot function for all APAF scripts}
\usage{
plotVolcano(
  exp,
  FC,
  pval,
  counts,
  PeporProt,
  comp.idx,
  FCCutoff = 1.5,
  pvalcutoff = 0.05
)
}
\arguments{
\item{exp}{SWATH or TMT}

\item{counts}{number of times target has been detected}

\item{PeporProt}{Specify 'Protein' or 'Peptide' depending on the plot. No other options are accepted}

\item{FCCutoff}{Fold Change cut off. Should be passed in from the script environment, but defaults to 1.5}

\item{ProtStats}{ProteinName, pval, FC}
}
\description{
The Volcano plot function for all APAF scripts
}
