% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcano.R
\name{plotVolcano}
\alias{plotVolcano}
\title{The Volcano plot function for all APAF scripts}
\usage{
plotVolcano(
  experiment = c("SWATH", "TMT"),
  FC,
  pval,
  PeporProt = c("Peptide", "Protein"),
  comp.idx,
  FCcutoff = 1.5,
  pvalcutoff = 0.05,
  saveInteractive = TRUE
)
}
\arguments{
\item{experiment}{SWATH or TMT}

\item{PeporProt}{Specify 'Protein' or 'Peptide' depending on the plot. No other options are accepted}

\item{FCcutoff}{Fold Change cut off. Should be passed in from the script environment, but defaults to 1.5}

\item{interactive}{TRUE or FALSE to plot an interactive volcano plot into a HTML file}
}
\description{
The Volcano plot function for all APAF scripts
}
