% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcano.R
\name{plotVolcano}
\alias{plotVolcano}
\title{The Volcano plot function for all APAF scripts}
\usage{
plotVolcano(
  ProtPepCounts = ag.pepN,
  ProtStats = ProtStats,
  PeporProt,
  FCcutoff = 1.5,
  pvalcutoff = 0.05
)
}
\arguments{
\item{ProtPepCounts}{ProteinName, NumberofPeptides}

\item{ProtStats}{ProteinName, pval, FC}

\item{PeporProt}{Specify 'Protein' or 'Peptide' depending on the plot. No other options are accepted}

\item{FCcutoff}{Fold Change cut off. Should be passed in from the script environment, but defaults to 1.5}

\item{pvalcutoff}{P-value cut off. Should be passed in from the script environment, but defaults to 0.05}
}
\description{
The Volcano plot function for all APAF scripts
}
